// src/types/fetchingTags.ts
export enum EFetchingTags {
    USER = 'USER',
    ME = 'ME',
    ACCOUNTS = 'ACCOUNTS',
    ACCOUNT = 'ACCOUNT',
    TRANSACTION_CATEGORIES = 'TRANSACTION_CATEGORIES',
    TRANSACTION_CATEGORY = 'TRANSACTION_CATEGORY',
    TRANSACTIONS = 'TRANSACTIONS',
    TRANSACTION = 'TRANSACTION',
    CALCULATED_TRANSACTION_VALUES = 'CALCULATED_TRANSACTION_VALUES',
}